requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyMuPDF==1.23.8
Pillow==10.1.0
ebooklib==0.18 